state manager observer and get unit tests
